+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Far faster implementation of `stats.coclustering`
 * Introducing `helpers.make_slurm_command`
+* Global submodule import (no longer need to import submodules individually)
 
 ## 0.1.0 (2020-03-06)